feat: Integrate deployed contracts with frontend
- Extract actual ABIs from deployed contracts
- Create comprehensive contract integration test component
- Add contract integration hooks for testing connections
- Create circle creation form with deployed contract integration
- Fix all TypeScript compilation issues
- Frontend now successfully builds with deployed contract addresses
- Ready for end-to-end testing on Base Sepolia

+++ frontend/src/hooks/useGaslessTransaction.ts
@@ -28,7 +28,7 @@ export function useGaslessTransaction() {
       return false;
     }
 
-    const eligible = isEligibleForGasless(trustScore, options.value);
+    const eligible = isEligibleForGasless(trustScore as bigint, options.value);
     if (!eligible) {
       setGaslessError(
         options.value 
@@ -88,7 +88,7 @@ export function useGaslessTransaction() {
       error: gaslessError,
       isGasless,
       userTrustScore: trustScore ? Number(trustScore) : 0,
-      maxValue: trustScore && trustScore >= BigInt(500) ? 1000 : 100, // $1000 or $100 USDC
+      maxValue: trustScore && (trustScore as bigint) >= BigInt(500) ? 1000 : 100, // $1000 or $100 USDC
     };
   }, [checkGaslessEligibility, gaslessError, isGasless, trustScore]);
